% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxreg.R
\name{coxreg}
\alias{coxreg}
\title{Cox Regression}
\usage{
coxreg(alpha, beta, P, B, RSquared, sigma)
}
\arguments{
\item{alpha}{type I error rate}

\item{beta}{1-power}

\item{P}{the proportion of subjects that fail}

\item{B}{Log Hazard Ratio}

\item{RSquared}{This is the R-Squared that is obtained when X1
is regressed on the other Xs (covariates) in the model}

\item{sigma}{Standard Deviation of X1}
}
\value{
N
}
\description{
Cox Regression
}
\examples{
coxreg(0.05, 0.2, 0.738, 1, 0.1837, 0.3126)
}
