% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrank.freedman.R
\name{logrank.freedman}
\alias{logrank.freedman}
\title{Logrank Tests (Freedman)}
\usage{
logrank.freedman(
  alpha,
  beta,
  k,
  pl,
  HRtype = c("hazard.rates", "survival.proportions", "median.survival.times"),
  param1,
  param2
)
}
\arguments{
\item{alpha}{type I error rate}

\item{beta}{1-power}

\item{k}{Group Allocation experimental group/control group}

\item{pl}{Proportion Lost During Follow Up}

\item{HRtype}{Survival Proportions}

\item{param1}{control group}

\item{param2}{experimental group}
}
\value{
N
}
\description{
Two Survival Curves,Test (Inequality)
}
\examples{
logrank.freedman(
  alpha = 0.05, beta = 1 - 0.70137, k = 1, pl = 0,
  HRtype = "survival.proportions", param1 = 0.5, param2 = 0.6
)
}
