% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestsforOneProportion.R
\name{TestsforOneProportion}
\alias{TestsforOneProportion}
\title{Title Tests for One Proportion}
\usage{
TestsforOneProportion(
  alternative = c("two", "upper", "lower"),
  beta,
  alpha,
  p0,
  p1
)
}
\arguments{
\item{alternative}{Two-Sided or One-side}

\item{beta}{1-power}

\item{alpha}{type I error rate}

\item{p0}{Null Proportion}

\item{p1}{Alternative Proportion}
}
\value{
n
}
\description{
Z Test using S(P0)
}
\examples{
TestsforOneProportion(alternative = "two", beta = 0.1, alpha = 0.05, p0 = 0.5, p1 = 0.55)
TestsforOneProportion(alternative = "upper", beta = 0.1, alpha = 0.05, p0 = 0.5, p1 = 0.55)
TestsforOneProportion(alternative = "lower", beta = 0.1, alpha = 0.05, p0 = 0.5, p1 = 0.45)
}
